Move factories to the same folder as the class they are creating (#99)
