+++ src/ConfigProvider.php
@@ -20,18 +20,18 @@ class ConfigProvider
         return [
             'factories' => [
                 /* Factories that do not map to a class */
-                'LmcRbacMvc\Guards' => \LmcRbacMvc\Factory\GuardsFactory::class,
+                'LmcRbacMvc\Guards' => \LmcRbacMvc\Guard\GuardsFactory::class,
 
                 /* Factories that map to a class */
-                \LmcRbacMvc\Assertion\AssertionPluginManager::class        => \LmcRbacMvc\Factory\AssertionPluginManagerFactory::class,
-                \LmcRbacMvc\Guard\GuardPluginManager::class                => \LmcRbacMvc\Factory\GuardPluginManagerFactory::class,
-                \LmcRbacMvc\Identity\AuthenticationIdentityProvider::class => \LmcRbacMvc\Factory\AuthenticationIdentityProviderFactory::class,
-                \LmcRbacMvc\Options\ModuleOptions::class                   => \LmcRbacMvc\Factory\ModuleOptionsFactory::class,
-                \LmcRbacMvc\Role\RoleProviderPluginManager::class          => \LmcRbacMvc\Factory\RoleProviderPluginManagerFactory::class,
-                \LmcRbacMvc\Service\AuthorizationService::class            => \LmcRbacMvc\Factory\AuthorizationServiceFactory::class,
-                \LmcRbacMvc\Service\RoleService::class                     => \LmcRbacMvc\Factory\RoleServiceFactory::class,
-                \LmcRbacMvc\View\Strategy\RedirectStrategy::class          => \LmcRbacMvc\Factory\RedirectStrategyFactory::class,
-                \LmcRbacMvc\View\Strategy\UnauthorizedStrategy::class      => \LmcRbacMvc\Factory\UnauthorizedStrategyFactory::class,
+                \LmcRbacMvc\Assertion\AssertionPluginManager::class        => \LmcRbacMvc\Assertion\AssertionPluginManagerFactory::class,
+                \LmcRbacMvc\Guard\GuardPluginManager::class                => \LmcRbacMvc\Guard\GuardPluginManagerFactory::class,
+                \LmcRbacMvc\Identity\AuthenticationIdentityProvider::class => \LmcRbacMvc\Identity\AuthenticationIdentityProviderFactory::class,
+                \LmcRbacMvc\Options\ModuleOptions::class                   => \LmcRbacMvc\Options\ModuleOptionsFactory::class,
+                \LmcRbacMvc\Role\RoleProviderPluginManager::class          => \LmcRbacMvc\Role\RoleProviderPluginManagerFactory::class,
+                \LmcRbacMvc\Service\AuthorizationService::class            => \LmcRbacMvc\Service\AuthorizationServiceFactory::class,
+                \LmcRbacMvc\Service\RoleService::class                     => \LmcRbacMvc\Service\RoleServiceFactory::class,
+                \LmcRbacMvc\View\Strategy\RedirectStrategy::class          => \LmcRbacMvc\View\Strategy\RedirectStrategyFactory::class,
+                \LmcRbacMvc\View\Strategy\UnauthorizedStrategy::class      => \LmcRbacMvc\View\Strategy\UnauthorizedStrategyFactory::class,
             ],
         ];
     }
@@ -54,7 +54,7 @@ class ConfigProvider
     {
         return [
             'factories' => [
-                \LmcRbacMvc\Mvc\Controller\Plugin\IsGranted::class => \LmcRbacMvc\Factory\IsGrantedPluginFactory::class,
+                \LmcRbacMvc\Mvc\Controller\Plugin\IsGranted::class => \LmcRbacMvc\Mvc\Controller\Plugin\IsGrantedPluginFactory::class,
             ],
             'aliases' => [
                 'isGranted' => \LmcRbacMvc\Mvc\Controller\Plugin\IsGranted::class,
@@ -66,8 +66,8 @@ class ConfigProvider
     {
         return [
             'factories' => [
-                \LmcRbacMvc\View\Helper\IsGranted::class => \LmcRbacMvc\Factory\IsGrantedViewHelperFactory::class,
-                \LmcRbacMvc\View\Helper\HasRole::class   => \LmcRbacMvc\Factory\HasRoleViewHelperFactory::class,
+                \LmcRbacMvc\View\Helper\IsGranted::class => \LmcRbacMvc\View\Helper\IsGrantedViewHelperFactory::class,
+                \LmcRbacMvc\View\Helper\HasRole::class   => \LmcRbacMvc\View\Helper\HasRoleViewHelperFactory::class,
             ],
             'aliases' => [
                 'isGranted' => \LmcRbacMvc\View\Helper\IsGranted::class,
