+++ src/Factory/ModuleOptionsFactory.php
@@ -18,29 +18,13 @@
 
 namespace LmcRbacMvc\Factory;
 
-use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
-use Psr\Container\ContainerInterface;
-use Laminas\ServiceManager\Factory\FactoryInterface;
-use Laminas\ServiceManager\ServiceLocatorInterface;
-use LmcRbacMvc\Options\ModuleOptions;
-
 /**
  * Factory for the module options
  *
  * @author  Michaël Gallego <mic.gallego@gmail.com>
  * @license MIT
+ * @deprecated Replaced by \LmcRbacMvc\Options\ModuleOptionsFactory
  */
-class ModuleOptionsFactory implements FactoryInterface
+class ModuleOptionsFactory extends \LmcRbacMvc\Options\ModuleOptionsFactory
 {
-    /**
-     * {@inheritDoc}
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): ModuleOptions
-    {
-        $config = $container->get('Config');
-        if (!isset($config['lmc_rbac'])) {
-            throw new ServiceNotCreatedException("Could not find the `lmc_rbac' configuration array ");
-        }
-        return new ModuleOptions($container->get('Config')['lmc_rbac']);
-    }
 }
