+++ src/Factory/ObjectRepositoryRoleProviderFactory.php
@@ -18,69 +18,13 @@
 
 namespace LmcRbacMvc\Factory;
 
-use Doctrine\Persistence\ObjectManager;
-use Doctrine\Persistence\ObjectRepository;
-use Psr\Container\ContainerInterface;
-use Laminas\ServiceManager\Factory\FactoryInterface;
-use Laminas\ServiceManager\ServiceLocatorInterface;
-use LmcRbacMvc\Exception;
-use LmcRbacMvc\Role\ObjectRepositoryRoleProvider;
-
 /**
  * Factory used to create an object repository role provider
  *
  * @author  Michaël Gallego <mic.gallego@gmail.com>
  * @license MIT
+ * @deprecated Replaced by \LmcRbacMvc\Role\ObjectRepositoryRoleProviderFactory
  */
-class ObjectRepositoryRoleProviderFactory implements FactoryInterface
+class ObjectRepositoryRoleProviderFactory extends \LmcRbacMvc\Role\ObjectRepositoryRoleProviderFactory
 {
-    protected array $options = [];
-
-    /**
-     * @param array $options
-     */
-    public function __construct(array $options = [])
-    {
-        $this->setCreationOptions($options);
-    }
-
-    /**
-     * @param array $options
-     */
-    public function setCreationOptions(array $options): void
-    {
-        $this->options = $options;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): ObjectRepositoryRoleProvider
-    {
-        $objectRepository = null;
-
-        if (!isset($options['role_name_property'])) {
-            throw new Exception\RuntimeException('The "role_name_property" option is missing');
-        }
-
-        if (isset($options['object_repository'])) {
-            /* @var ObjectRepository $objectRepository */
-            $objectRepository = $container->get($options['object_repository']);
-
-            return new ObjectRepositoryRoleProvider($objectRepository, $options['role_name_property']);
-        }
-
-        if (isset($options['object_manager']) && isset($options['class_name'])) {
-            /* @var ObjectManager $objectManager */
-            $objectManager    = $container->get($options['object_manager']);
-            $objectRepository = $objectManager->getRepository($options['class_name']);
-
-            return new ObjectRepositoryRoleProvider($objectRepository, $options['role_name_property']);
-        }
-
-        throw new Exception\RuntimeException(
-            'No object repository was found while creating the LmcRbacMvc object repository role provider. Are
-             you sure you specified either the "object_repository" option or "object_manager"/"class_name" options?'
-        );
-    }
 }
