+++ src/Factory/RoutePermissionsGuardFactory.php
@@ -18,59 +18,14 @@
 
 namespace LmcRbacMvc\Factory;
 
-use Psr\Container\ContainerInterface;
-use Laminas\ServiceManager\Factory\FactoryInterface;
-use Laminas\ServiceManager\ServiceLocatorInterface;
-use LmcRbacMvc\Guard\RouteGuard;
-use LmcRbacMvc\Guard\RoutePermissionsGuard;
-use LmcRbacMvc\Options\ModuleOptions;
-use LmcRbacMvc\Service\AuthorizationService;
-
 /**
  * Create a route guard for checking permissions
  *
  * @author  Michaël Gallego <mic.gallego@gmail.com>
  * @author  JM Lerouxw <jmleroux.pro@gmail.com>
  * @license MIT
+ * @deprecated Replaced by \LmcRbacMvc\Guard\RoutePermissionsGuardFactory
  */
-class RoutePermissionsGuardFactory implements FactoryInterface
+class RoutePermissionsGuardFactory extends \LmcRbacMvc\Guard\RoutePermissionsGuardFactory
 {
-    protected array $options = [];
-
-    /**
-     * @param array $options
-     */
-    public function __construct(array $options = [])
-    {
-        $this->setCreationOptions($options);
-    }
-
-    /**
-     * @param array $options
-     */
-    public function setCreationOptions(array $options): void
-    {
-        $this->options = $options;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): RoutePermissionsGuard
-    {
-        if (null === $options) {
-            $options = [];
-        }
-
-        /* @var ModuleOptions $moduleOptions */
-        $moduleOptions = $container->get(ModuleOptions::class);
-
-        /* @var AuthorizationService $authorizationService */
-        $authorizationService = $container->get(AuthorizationService::class);
-
-        $routeGuard = new RoutePermissionsGuard($authorizationService, $options);
-        $routeGuard->setProtectionPolicy($moduleOptions->getProtectionPolicy());
-
-        return $routeGuard;
-    }
 }
