+++ src/Service/AuthorizationServiceDelegatorFactory.php
@@ -1,0 +1,58 @@
+<?php
+/*
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ * This software consists of voluntary contributions made by many individuals
+ * and is licensed under the MIT license.
+ */
+
+namespace LmcRbacMvc\Service;
+
+use Psr\Container\ContainerInterface;
+//use Laminas\ServiceManager\AbstractPluginManager;
+use Laminas\ServiceManager\Factory\DelegatorFactoryInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
+use LmcRbac\Exception\RuntimeException;
+use LmcRbacMvc\Service\AuthorizationService;
+use LmcRbacMvc\Service\AuthorizationServiceAwareInterface;
+
+/**
+ * Delegator factory for classes implementing AuthorizationServiceAwareInterface
+ *
+ * @author  Jean-Marie Leroux <jmleroux.pro@gmail.com>
+ * @license MIT License
+ */
+class AuthorizationServiceDelegatorFactory implements DelegatorFactoryInterface
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function __invoke(ContainerInterface $container, $name, callable $callback, array $options = null): AuthorizationServiceAwareInterface
+    {
+        $instanceToDecorate = call_user_func($callback);
+
+        if (!$instanceToDecorate instanceof AuthorizationServiceAwareInterface) {
+            throw new RuntimeException("The service $name must implement AuthorizationServiceAwareInterface.");
+        }
+        $authorizationService = $container->get(AuthorizationService::class);
+        $instanceToDecorate->setAuthorizationService($authorizationService);
+
+        return $instanceToDecorate;
+    }
+
+    /** TODO remove */
+    public function createDelegatorWithName(ServiceLocatorInterface $serviceLocator, $name, $requestedName, $callback)
+    {
+        return $this($serviceLocator, $requestedName, $callback);
+    }
+}
