+++ tests/Guard/ControllerGuardFactoryTest.php
@@ -16,18 +16,18 @@
  * and is licensed under the MIT license.
  */
 
-namespace LmcRbacMvcTest\Factory;
+namespace LmcRbacMvcTest\Guard;
 
 use Laminas\ServiceManager\ServiceManager;
-use LmcRbacMvc\Factory\RoutePermissionsGuardFactory;
+use LmcRbacMvc\Guard\ControllerGuardFactory;
 use LmcRbacMvc\Guard\GuardInterface;
 use LmcRbacMvc\Guard\GuardPluginManager;
 use LmcRbacMvc\Options\ModuleOptions;
 
 /**
- * @covers \LmcRbacMvc\Factory\RoutePermissionsGuardFactory
+ * @covers \LmcRbacMvc\Guard\ControllerGuardFactory
  */
-class RoutePermissionsGuardFactoryTest extends \PHPUnit\Framework\TestCase
+class ControllerGuardFactoryTest extends \PHPUnit\Framework\TestCase
 {
     public function testFactory()
     {
@@ -37,28 +37,28 @@ class RoutePermissionsGuardFactoryTest extends \PHPUnit\Framework\TestCase
             $this->markTestSkipped('this test is only vor zend-servicemanager v3');
         }
 
-        $creationOptions = [
-            'route' => 'role'
-        ];
-
         $options = new ModuleOptions([
             'identity_provider' => 'LmcRbacMvc\Identity\AuthenticationProvider',
             'guards'            => [
-                'LmcRbacMvc\Guard\RoutePermissionsGuard' => $creationOptions
+                'LmcRbacMvc\Guard\ControllerGuard' => [
+                    'controller' => 'MyController',
+                    'actions'    => 'edit',
+                    'roles'      => 'member'
+                ]
             ],
-            'protection_policy' => GuardInterface::POLICY_ALLOW,
+            'protection_policy' => GuardInterface::POLICY_ALLOW
         ]);
 
         $serviceManager->setService('LmcRbacMvc\Options\ModuleOptions', $options);
         $serviceManager->setService(
-            'LmcRbacMvc\Service\AuthorizationService',
-            $this->getMockBuilder('LmcRbacMvc\Service\AuthorizationService')->disableOriginalConstructor()->getMock()
+            'LmcRbacMvc\Service\RoleService',
+            $this->getMockBuilder('LmcRbacMvc\Service\RoleService')->disableOriginalConstructor()->getMock()
         );
 
-        $factory    = new RoutePermissionsGuardFactory();
-        $routeGuard = $factory($serviceManager, 'LmcRbacMvc\Guard\RoutePermissionsGuard');
+        $factory         = new ControllerGuardFactory();
+        $controllerGuard = $factory($serviceManager, GuardPluginManager::class);
 
-        $this->assertInstanceOf('LmcRbacMvc\Guard\RoutePermissionsGuard', $routeGuard);
-        $this->assertEquals(GuardInterface::POLICY_ALLOW, $routeGuard->getProtectionPolicy());
+        $this->assertInstanceOf('LmcRbacMvc\Guard\ControllerGuard', $controllerGuard);
+        $this->assertEquals(GuardInterface::POLICY_ALLOW, $controllerGuard->getProtectionPolicy());
     }
 }
